test: rearrange test directories
