+++ package.json
@@ -87,19 +87,19 @@
     "webpack-cli": "^3.3.4"
   },
   "scripts": {
-    "codecov": "nyc mocha build/test --reporter spec --slow 500 && codecov",
+    "codecov": "nyc mocha build/test/unit --reporter spec --slow 500 && codecov",
     "docs": "compodoc src/",
     "gen-parser": "pegjs src/pathTemplateParser.pegjs",
     "pretest": "npm run prepare",
-    "test": "nyc mocha build/test",
+    "test": "nyc mocha build/test/unit",
     "lint": "gts check && eslint samples/*.js samples/**/*.js",
     "clean": "gts clean",
-    "compile": "tsc -p . && cp src/*.json build/src && cp src/*.js build/src && cp -r system-test/fixtures build/system-test/",
+    "compile": "tsc -p . && cp src/*.json build/src && cp src/*.js build/src && cp -r test/fixtures build/test/",
     "compile-protos": "pbjs -t json ./protos/google/longrunning/operations.proto -p ./protos > protos/operations.json",
     "fix": "gts fix && eslint --fix samples/*.js samples/**/*.js",
     "prepare": "npm run compile && node ./build/tools/prepublish.js",
     "posttest": "npm run lint",
-    "system-test": "nyc mocha build/system-test/test.*.js --timeout 120000",
+    "system-test": "nyc mocha build/test/system-test --timeout 120000",
     "samples-test": "echo no sample tests 😱",
     "docs-test": "linkinator docs",
     "predocs-test": "npm run docs",
