+++ test/fixtures/google-gax-packaging-test-app/protos/google/showcase/v1beta1/identity.proto
