+++ test/fixtures/google-gax-packaging-test-app/src/v1beta1/echo_client_config.json
