+++ test/unit/compileProtos.ts
@@ -36,7 +36,7 @@ import * as rimraf from 'rimraf';
 import * as util from 'util';
 import * as path from 'path';
 import * as protobuf from 'protobufjs';
-import * as compileProtos from '../tools/compileProtos';
+import * as compileProtos from '../../tools/compileProtos';
 
 const readFile = util.promisify(fs.readFile);
 const mkdir = util.promisify(fs.mkdir);
