+++ test/unit/gax.ts
@@ -35,7 +35,7 @@
  * Therefore quote-props is disabled explicitly only in this file. */
 
 import {expect} from 'chai';
-import * as gax from '../src/gax';
+import * as gax from '../../src/gax';
 
 const SERVICE_NAME = 'test.interface.v1.api';
